zaakafhandelcomponent-2168 documentcomponent icoon + standardisatie
