+++ src/main/app/src/app/zaken/zaak-documenten/zaak-documenten.component.less
@@ -11,3 +11,23 @@
 #downloadZip_button:hover {
   color: #000000;
 }
+
+.blue {
+  color: blue;
+}
+
+.green {
+  color: green;
+}
+
+.orange {
+  color: orange;
+}
+
+.red {
+  color: red;
+}
+
+.darkred {
+  color: darkred;
+}
