+++ src/main/app/src/assets/i18n/en.json
@@ -93,6 +93,7 @@
   "aangemaaktDoor": "Created by",
   "afgerond": "Finished",
   "behandelaar": "Case handler",
+  "bestand": "File",
   "bestandsnaam": "File name",
   "achternaam": "Surname",
   "voorvoegsel": "Prefix",
