+++ src/main/app/src/assets/i18n/nl.json
@@ -93,6 +93,7 @@
   "afgerond": "Afgerond",
   "auteur": "Auteur",
   "behandelaar": "Behandelaar",
+  "bestand": "Bestand",
   "bestandsnaam": "Bestandsnaam",
   "achternaam": "Achternaam",
   "voorvoegsel": "Voorvoegsel",
